feature: customize the PML to account for beam path in non vacuum material (#34)
* fix PML conductivity tapering function for X axis

* remove whitespaces in functions call

* Pass PML parameters as input to SolverFIT3D

Besides the number of PML cells, the resistivity boundaries, permittivity and the function used to compute the resistivity profile in the PML are now arguments for the SolverFIT3D

* Change default values for PML resistivity and permittivity

* Add docstring for PML parameters

* remove PML parameters from SolverFIT3D init arguments

To make the init clearer, the PML parameters are given default values within the initialization. To customize the values for PML resistivity ans eps, user can access the attributes and re-execute fill_pml_sigmas to update the PML cells.

* Use adjacent layer properties for PML conductivity and permittivity

This first attempt uses the properties of the central cell of the layer adjacent to the PML, assuming the beam is centered

* fix: PML conductivity calculation
Relaxation times were incorrect when sigma=0 in the simulation space, as the whole PML had sigma=0

* fix: avoid null sigma in PML while guaranteeing increase in sigma for cases other than vacuum

The sigma for the PML layers is now the sigma of the previous non-PML layer + the function of resistivity. This function starts from the sigma value of the layer if it is non zero, or from a low sigma value

* fix: update PML conductivity limits

* fix: update PML high conductivity limit to 10.0
This is to improve field absorption, in particular if we have vacuum. One might need more layers of PML then

* fix: update sigma calculation to avoid null values in PML for relaxation time computation

* test: add some tests for resistive materials with PML

This still needs to be worked out, as there are some divergence in computations

* fix: update eps_r and sigma values for PML boundary tests